S3WithGlacier/requirements.txt
-e .
aws-cdk.aws_s3
